Add TestRunFilter custom marshalling
